extract utils.js to separate file from init/init-repeater.js
