+++ generator/data/contract-data.js
@@ -7,10 +7,10 @@ export const contractData = {
         name: "",
         email: ""
     },
-    // shortcodes: []
     shortcodes: [
         { name: '[price]', value: '$1000' },
         { name: '[date]', value: '31/12/2024' },
         { name: '[hourly-rate]', value: '$50' },
-    ]
+    ],
+    // shortcodes: []
 }
