+++ generator/scripts/init/init-repeater.js
@@ -1,65 +1,5 @@
 import { contractData } from "../../data/contract-data.js"
-// export const contractData = {
-//     client: {
-//         name: "",
-//         email: ""
-//     },
-//     dev: {
-//         name: "",
-//         email: ""
-//     },
-//     // shortcodes: [],
-//     shortcodes: [
-//         { name: '[price]', value: '$1000' },
-//         { name: '[date]', value: '31/12/2024' },
-//         { name: '[hourly-rate]', value: '$50' },
-//     ]
-// };
-// import { select, addEvent, debounce, addBrackets } from "../utils.js"
-export const utils = {
-    // wrapper around querySelector
-    // select() usage: const [sel1, sel2] = select(".class1", ".class2")
-    // [repeaterForm, repeaterOutput, addItemButton] = select( ".repeater-form", "output", ".add-btn");
-    select: (...selectors) => selectors.map(selector => document.querySelector(selector)),
-
-    // selectAll() usage: const [sel1, sel2] = selectAll(".class1", ".class2")
-    // wrapper around querySelectorAll
-    selectAll: (...selectors) => selectors.map(selector => document.querySelectorAll(selector)),
-
-    // addEvent() usage: addEvent(element, handler, eventType)
-    // wrapper around addEventListener with optional event type at the end
-    addEvent: (element, handler, eventType = "click") => element.addEventListener(eventType, handler),
-
-    // debounce() usage: const debouncedMyFunc = debounce(myFunc, 200);
-    // or function debouncedMyFunc() { return debounce(myFunc, 200); }
-    debounce: (callback, wait = 200) => {
-        let timeoutId = null;
-        return function (...args) {
-            window.clearTimeout(timeoutId);
-            timeoutId = window.setTimeout(() => {
-                callback.apply(null, args);
-            }, wait);
-        };
-    },
-
-    // addBrackets() usage: addBrackets(element)
-    addBrackets: (element) => {
-        let trimmedValue = element.value.trim();
-
-        if (!trimmedValue.startsWith('[')) {
-            trimmedValue = '[' + trimmedValue;
-        }
-
-        if (!trimmedValue.endsWith(']')) {
-            trimmedValue = trimmedValue + ']';
-        }
-
-        element.value = trimmedValue;
-    },
-};
-const { addEvent, debounce, addBrackets } = { ...utils };
-
-
+import { addEvent, debounce } from "../utils.js"
 
 export default function repeater(repeater = ".repeater-form", output = "output", addBtn = ".add-btn") {
     // Cache frequently used elements
